Clean up I650 interlock loop sign comparison
Cast the fixed InterLockCount array bound to the existing signed loop
index type. The loop walks a small local interlock table, so this clears
-Wsign-compare without changing behavior.

+++ simulators/I650/i650_cpu.c
@@ -1800,7 +1800,7 @@ sim_instr(void)
         GlobalWordTimeCount++;
 
         // if any interlock set, decrease it
-        for (il=0;il < sizeof(InterLockCount)/ sizeof(InterLockCount[0]) ;il++) {
+        for (il=0;il < (int)(sizeof(InterLockCount)/ sizeof(InterLockCount[0])) ;il++) {
             if (InterLockCount[il] > 0) InterLockCount[il]--;
         }
         // decrease pending to execute step intruction count
